Refactor : video-component
